feat: try polygon dividing
