+++ index.js
@@ -2,4 +2,3 @@ export * from './lib/zrender';
 export * from './lib/export';
 
 import './lib/svg/svg';
-import './lib/vml/vml';
