+++ src/tool/dividePath.ts
@@ -1,13 +1,17 @@
+import { fromPoints } from '../core/bbox';
+import BoundingRect, { RectLike } from '../core/BoundingRect';
+import { Point } from '../export';
 import Path from '../graphic/Path';
 import Polygon from '../graphic/shape/Polygon';
 import Rect from '../graphic/shape/Rect';
 import Sector from '../graphic/shape/Sector';
+import { pathToPolygons } from './convertPath';
 import { clonePath } from './path';
 
 // Default shape dividers
 
 interface BinaryDivide {
-    (shape: Path['shape'], out: Path['shape'][]): void
+    (shape: Path['shape']): Path['shape'][]
 }
 
 const SECTOR_COMMON_PROPS: (keyof Sector['shape'])[] = ['clockwise', 'cornerRadius', 'innerCornerRadius', 'cx', 'cy'];
@@ -22,7 +26,7 @@ function copyShapeProps(out: Path['shape'][], source: Path['shape'], keys: strin
     }
 }
 
-function binaryDivideSector(sectorShape: Sector['shape'], out: Sector['shape'][]) {
+function binaryDivideSector(sectorShape: Sector['shape']) {
     // Divide into two
     const r0 = sectorShape.r0;
     const r = sectorShape.r;
@@ -30,6 +34,7 @@ function binaryDivideSector(sectorShape: Sector['shape'], out: Sector['shape'][]
     const endAngle = sectorShape.endAngle;
     const angle = Math.abs(endAngle - startAngle);
     const arcLen = angle * r;
+    const out: Sector['shape'][] = [];
     if (arcLen < Math.abs(r - r0)) {
         const midR = (r0 + r) / 2;
         // Divide on radius
@@ -64,15 +69,18 @@ function binaryDivideSector(sectorShape: Sector['shape'], out: Sector['shape'][]
     }
 
     copyShapeProps(out, sectorShape, SECTOR_COMMON_PROPS);
+
+    return out;
 }
 
 
-function binaryDivideRect(rectShape: Rect['shape'], out: Rect['shape'][]) {
+function binaryDivideRect(rectShape: Rect['shape']) {
     const width = rectShape.width;
     const height = rectShape.height;
     const x = rectShape.x;
     const y = rectShape.y;
 
+    const out: Rect['shape'][] = [];
     if (width < height) {
         const halfHeight = height / 2;
         out[0] = {
@@ -99,10 +107,142 @@ function binaryDivideRect(rectShape: Rect['shape'], out: Rect['shape'][]) {
     if (rectShape.r != null) {
         out[0].r = out[1].r = rectShape.r;
     }
+    return out;
 }
 
-function binaryDividePolygon(poygonShape: Polygon['shape'], out: Polygon['shape'][]) {
+function crossProduct2d(x1: number, y1: number, x2: number, y2: number) {
+    return x1 * y2 - x2 * y1;
+}
 
+function lineLineIntersect(
+    a1x: number, a1y: number, a2x: number, a2y: number, // p1
+    b1x: number, b1y: number, b2x: number, b2y: number // p2
+): Point {
+    const mx = a2x - a1x;
+    const my = a2y - a1y;
+    const nx = b2x - b1x;
+    const ny = b2y - b1y;
+
+    const nmCrossProduct = crossProduct2d(nx, ny, mx, my);
+    if (Math.abs(nmCrossProduct) < 1e-6) {
+        return null;
+    }
+
+    const b1a1x = a1x - b1x;
+    const b1a1y = a1y - b1y;
+
+    const q = crossProduct2d(b1a1x, b1a1y, mx, my) / nmCrossProduct;
+    if (q < 0 || q > 1) {
+        return null;
+    }
+    // p2 is an infinite line
+    return new Point(
+        q * mx + a1x,
+        q * my + a1y
+    );
+}
+
+function projPtOnLine(pt: Point, lineA: Point, lineB: Point): Point {
+    const dir = new Point();
+    Point.sub(dir, lineB, lineA);
+    dir.normalize();
+    const dir2 = new Point();
+    Point.sub(dir2, pt, lineA);
+    const len = dir2.dot(dir);
+    Point.scaleAndAdd(dir, lineA, dir, len);
+    return dir;
+}
+
+function splitPolygonByLine(points: number[][], lineA: Point, lineB: Point) {
+    const len = points.length;
+    const intersections: {
+        projPt: Point,
+        pt: Point
+        idx: number
+    }[] = [];
+    for (let i = 0; i < len; i++) {
+        const p0 = points[i];
+        const p1 = points[(i + 1) % len];
+        const intersectionPt = lineLineIntersect(
+            p0[0], p0[1], p1[0], p1[1],
+            lineA.x, lineA.y, lineB.x, lineB.y
+        );
+        if (intersectionPt) {
+            intersections.push({
+                projPt: projPtOnLine(intersectionPt, lineA, lineB),
+                pt: intersectionPt,
+                idx: i
+            });
+        }
+    }
+
+    // TODO No intersection?
+
+    // Find two farthest points.
+    intersections.sort((a, b) => {
+        return a.projPt.x - b.projPt.x;
+    });
+    let splitPt0 = intersections[0];
+    let splitPt1 = intersections[intersections.length - 1];
+    if (splitPt1.idx < splitPt0.idx) {
+        const tmp = splitPt0;
+        splitPt0 = splitPt1;
+        splitPt1 = tmp;
+    }
+
+    const newPolyA: number[][] = [
+        [splitPt0.pt.x, splitPt0.pt.y]
+    ];
+    const newPolyB: number[][] = [
+        [splitPt1.pt.x, splitPt1.pt.y]
+    ];
+
+    for (let i = splitPt0.idx + 1; i < splitPt1.idx; i++) {
+        newPolyA.push(points[i].slice());
+    }
+    newPolyA.push([splitPt1.pt.x, splitPt1.pt.y]);
+
+    for (let i = splitPt1.idx + 1; i < splitPt1.idx + len; i++) {
+        newPolyB.push(points[i % len].slice());
+    }
+    newPolyB.push([splitPt1.pt.x, splitPt1.pt.y]);
+
+    return [{
+        points: newPolyA
+    }, {
+        points: newPolyB
+    }];
+}
+
+function binaryDividePolygon(
+    polygonShape: Pick<Polygon['shape'], 'points'>
+) {
+    const points = polygonShape.points;
+    const min: number[] = [];
+    const max: number[] = [];
+    fromPoints(points, min, max);
+    const boundingRect = new BoundingRect(
+        min[0], min[1], max[0] - min[0], max[1] - min[1]
+    );
+
+    const width = boundingRect.width;
+    const height = boundingRect.height;
+    const x = boundingRect.x;
+    const y = boundingRect.y;
+
+    const pt0 = new Point();
+    const pt1 = new Point();
+    if (width > height) {
+        pt0.x = pt1.x = x + width / 2;
+        pt0.y = y;
+        pt1.y = y + height;
+    }
+    else {
+        pt0.y = pt1.y = y + height / 2;
+        pt0.x = x;
+        pt1.x = x + width;
+    }
+    return splitPolygonByLine(points, pt0, pt1);
 }
 
 
@@ -114,10 +254,9 @@ function binaryDivideRecursive<T extends Path['shape']>(
     }
     else {
         const mid = Math.floor(count / 2);
-        const tmpArr: Path['shape'][] = [];
-        divider(shape, tmpArr);
-        binaryDivideRecursive(divider, tmpArr[0], mid, out);
-        binaryDivideRecursive(divider, tmpArr[1], count - mid, out);
+        const sub = divider(shape);
+        binaryDivideRecursive(divider, sub[0], mid, out);
+        binaryDivideRecursive(divider, sub[1], count - mid, out);
     }
 
     return out;
@@ -136,6 +275,14 @@ function copyPathProps(source: Path, target: Path) {
     target.z = source.z;
     target.z2 = source.z2;
     target.zlevel = source.zlevel;
+}
+
+function polygonConvert(points: number[]): number[][] {
+    const out = [];
+    for (let i = 0; i < points.length;) {
+        out.push([points[i++], points[i++]]);
+    }
+    return out;
 }
 
 export function split(
@@ -160,6 +307,14 @@ export function split(
                 cx: shape.cx, cy: shape.cy
             } as Sector['shape'], count, outShapes);
             OutShapeCtor = Sector;
+            break;
+        default:
+            // TODO
+            const polygons = pathToPolygons(path.getUpdatedPathProxy());
+            binaryDivideRecursive(binaryDividePolygon, {
+                points: polygonConvert(polygons[0])
+            }, count, outShapes);
+            OutShapeCtor = Polygon;
             break;
     }
