+++ test/pathToPolygon.html
@@ -19,7 +19,7 @@
     <script>
         requireES([
             "zrender/lib/core/PathProxy",
-            "zrender/lib/tool/morphPath"
+            "zrender/lib/tool/convertPath"
         ], function(PathProxy, {pathToPolygons}) {
 
             const canvas = document.querySelector('#main');
